+++ src/test/java/bridge/domain/BridgeMakerTest.java
@@ -1,7 +1,7 @@
 //package bridge.domain;
 //
 //import bridge.BridgeRandomNumberGenerator;
-//import bridge.domain.generator.BridgeMaker;
+//import bridge.BridgeMaker;
 //import org.junit.jupiter.api.BeforeAll;
 //import org.junit.jupiter.api.DisplayName;
 //import org.junit.jupiter.api.Test;
